file navigator basic NO DATA default added
