+++ src/app/(post-login)/submission/program/[shortName]/clinical-submission/types.ts
@@ -1,4 +1,9 @@
 import { ClinicalSubmissionQuery } from "@/__generated__/graphql";
 
+export type ClinicalSubmission = ClinicalSubmissionQuery["clinicalSubmissions"];
+
+export type ClinicalSubmissionState =
+  ClinicalSubmissionQuery["clinicalSubmissions"]["state"];
+
 export type ClinicalEntities =
   ClinicalSubmissionQuery["clinicalSubmissions"]["clinicalEntities"];
